chat_interface/pyproject.toml
[tool.poetry]
name = "chat_interface"
version = "0.1.0"
authors = ["Bruno Freitas <bruno.freitas@neuralmind.ai>"]
description = "Chat interface for the Retail-GPT chatbot"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.31.1"
typing-extensions = "^4.9.0"
python-docx = "^1.1.0"
redis = "^5.0.7"
pillow = "^10.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.1"
pre-commit = "^3.6.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
